Reorganize scripts by folder
